Move `teleport integration configure` commands to their own file (#40634)
* Move teleport integration configure commands to their own file

* remove pointers for requests
